+++ CHANGES.md
@@ -11,10 +11,10 @@ Version 2.0.0 (MMM DD, 20yy)
      - methods are async wherever possible
  - where possible, @property is used instead of a simple getter method
  - new feature: new Job properties: exit_code, log
- - code cleanup
- - remove CWL code, remove connection helper
  - use httpx instead of requests as underlying HTTP library for both
    sync and async APIs
+ - code cleanup
+ - remove CWL code, remove connection helper
 
 Version 1.3.7 (Mar 17, 2026)
 ----------------------------