+++ CHANGES.md
@@ -3,6 +3,11 @@ Changelog for PyUNICORE
 
 Issue tracker: https://github.com/HumanBrainProject/pyunicore
 
+Version 1.3.8 (MMM DD, 2026)
+----------------------------
+ - CLI: exec / run use any site from registry when no sitename or
+   url is given
+
 Version 1.3.7 (Mar 17, 2026)
 ----------------------------
  - add helper code for uftp authserver